Server/requirements.txt
opencv-python
numpy
networkx
flask
pyyaml